[LMP BS] Fixed OS X build.

+++ src/plugins/lmp/plugins/brainslugz/checker.cpp
@@ -29,6 +29,7 @@
 
 #include "checker.h"
 #include <algorithm>
+#include <cmath>
 #include <QtDebug>
 #include <interfaces/core/ipluginsmanager.h>
 #include <interfaces/lmp/ilmpproxy.h>
@@ -86,7 +87,7 @@ namespace BrainSlugz
 			const auto pos = std::find_if (releases.begin (), releases.end (),
 					[&albumPtr] (const Media::ReleaseInfo& release)
 					{
-						return std::abs (albumPtr->Year_ - release.Year_) <= 1 &&
+						return std::abs (static_cast<double> (albumPtr->Year_ - release.Year_)) <= 1 &&
 								!QString::compare (albumPtr->Name_, release.Name_, Qt::CaseInsensitive);
 					});
